removed main thread from GeodeRedisServer
removed NamedThreadFactory in favor of geode's LoggingThreadFactory.
Also now uses LoggingExecutors instead of Executors.
These classes makes sure that any unhandled exceptions that cause
a thread to die will be logged.

removed logLevel since it is configured by gemfire

A cache is now passed to the GeodeRedisServer constructor

region initialization now done in RegionProvider

extracted passive expiration code

extracted netty server code
also reduced dependencies on GeodeRedisServer

renamed close to stop

improved exception handling and cleanup up NetyRedisServer

changed passive expiration to use scheduleWithFixedDelay

fixed bind exception handling
