+++ geode-redis/src/main/java/org/apache/geode/redis/internal/GeodeRedisService.java
@@ -14,6 +14,7 @@
  */
 package org.apache.geode.redis.internal;
 
+
 import org.apache.logging.log4j.Logger;
 
 import org.apache.geode.cache.Cache;
@@ -77,7 +78,7 @@ public class GeodeRedisService implements CacheService, ResourceEventsListener {
           String.format("Starting GeodeRedisServer on bind address %s on port %s",
               new Object[] {bindAddress, port}));
 
-      this.redisServer = new GeodeRedisServer(bindAddress, port);
+      this.redisServer = new GeodeRedisServer(bindAddress, port, cache);
       this.redisServer.start();
     }
   }
