--- geode-redis/src/main/java/org/apache/geode/redis/internal/NamedThreadFactory.java
@@ -1,38 +1,0 @@
-/*
- * Licensed to the Apache Software Foundation (ASF) under one or more contributor license
- * agreements. See the NOTICE file distributed with this work for additional information regarding
- * copyright ownership. The ASF licenses this file to You under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance with the License. You may obtain a
- * copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software distributed under the License
- * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
- * or implied. See the License for the specific language governing permissions and limitations under
- * the License.
- *
- */
-package org.apache.geode.redis.internal;
-
-import java.util.concurrent.ThreadFactory;
-import java.util.concurrent.atomic.AtomicInteger;
-
-public class NamedThreadFactory implements ThreadFactory {
-  private final AtomicInteger counter = new AtomicInteger();
-  private final String baseName;
-  private final boolean isDaemon;
-
-  public NamedThreadFactory(String baseName, boolean isDaemon) {
-    this.baseName = baseName;
-    this.isDaemon = isDaemon;
-  }
-
-  @Override
-  public Thread newThread(Runnable runnable) {
-    Thread thread = new Thread(runnable);
-    thread.setName(baseName + counter.incrementAndGet());
-    thread.setDaemon(isDaemon);
-    return thread;
-  }
-}
